fix(plugin): add heighReference to gtfs bus models
This commit fixes the following \n1) HeightReference issue we are facing with gtfs Bus Model without chaning the how the rest of models(glTF) are handled. \n2)It also adds a refactor the merging method used for handling defaultOverrides with overrides, and named `mergeDefaultOverrides` at `plugin/src/utils/merge.ts`
